Rearchitect OnTypeFormattingProvider for extensibility
