+++ packages/theme-language-server-common/src/server/startServer.ts
@@ -204,7 +204,7 @@ export function startServer(
         },
         documentOnTypeFormattingProvider: {
           firstTriggerCharacter: ' ',
-          moreTriggerCharacter: ['{', '%', '-'],
+          moreTriggerCharacter: ['{', '%', '-', '>'],
         },
         documentLinkProvider: {
           resolveProvider: false,
